Fix play/pause clicks for media
...by tracking media state globally. It's lazy, but it works well
enough. Eventually I'll find a better way, but in the meantime, this
fixes a stupid annoying bug.
